Extract IssueTree to use in Epic
